requirements.txt
imgui[glfw]
glfw
PyOpenGL
numpy
pywin32
dxcam
av
pyaudio